# backend/openapi.json
{
  "components": {
    "schemas": {
      "BayDetailOut": {
        "properties": {
          "bay_type": {
            "title": "Bay Type",
            "type": "string"
          },
          "device_count": {
            "title": "Device Count",
            "type": "integer"
          },
          "devices": {
            "items": {
              "$ref": "#/components/schemas/DeviceOut"
            },
            "title": "Devices",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_live": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Live"
          },
          "is_live_quality": {
            "$ref": "#/components/schemas/Quality"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssueOut"
            },
            "title": "Issues",
            "type": "array"
          },
          "logical_nodes": {
            "items": {
              "type": "string"
            },
            "title": "Logical Nodes",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "voltage_level": {
            "title": "Voltage Level",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "voltage_level",
          "bay_type",
          "template_id",
          "device_count",
          "is_live",
          "is_live_quality",
          "logical_nodes",
          "issues",
          "devices"
        ],
        "title": "BayDetailOut",
        "type": "object"
      },
...
      "DeviceOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "ln": {
            "title": "Ln",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "quality": {
            "$ref": "#/components/schemas/Quality"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "short_name": {
            "title": "Short Name",
            "type": "string"
          },
          "source_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref"
          },
          "source_timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Timestamp"
          },
          "state": {
            "$ref": "#/components/schemas/SwitchState"
          },
          "terminals": {
            "items": {
              "type": "string"
            },
            "title": "Terminals",
            "type": "array"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "integer"
              },
              {
                "type": "boolean"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "id",
          "ln",
          "role",
          "name",
          "short_name",
          "state",
          "quality",
          "value",
          "source_timestamp",
          "source_ref",
          "terminals"
        ],
        "title": "DeviceOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginIn": {
        "properties": {
          "password": {
            "title": "Password",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginIn",
        "type": "object"
      },
      "MeOut": {
        "description": "The caller's identity and permissions (ADR-0016, ADR-0017).\n\n`capabilities` is the field the UI switches on. `roles` is for display and\nfor support calls — \"which roles is this account holding\" is the first\nquestion when somebody reports a missing button.\n\n`authenticated=False` means nobody is signed in, which is not an error: it\nis what `/api/me` answers before the login screen has been used.",
        "properties": {
          "auth_mode": {
            "default": "session",
            "title": "Auth Mode",
            "type": "string"
          },
          "authenticated": {
            "default": false,
            "title": "Authenticated",
            "type": "boolean"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "display_name": {
            "default": "",
            "title": "Display Name",
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "user": {
            "title": "User",
            "type": "string"
          }
        },
        "required": [
          "user",
          "roles",
          "capabilities"
        ],
        "title": "MeOut",
        "type": "object"
      },
...
      "PasswordChangeIn": {
        "properties": {
          "current_password": {
            "title": "Current Password",
            "type": "string"
          },
          "new_password": {
            "title": "New Password",
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "PasswordChangeIn",
        "type": "object"
      },
...
      "ProjectCreateIn": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "opcua_url": {
            "title": "Opcua Url",
            "type": "string"
          }
        },
        "required": [
          "name",
          "opcua_url"
        ],
        "title": "ProjectCreateIn",
        "type": "object"
      },
      "ProjectLoadOut": {
        "description": "Result of creating/opening/refreshing a project.\n\n`ok=False` means the project exists but its source could not be read —\nthe row is kept so the operator can fix the URL or the network and retry.",
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectOut"
          }
        },
        "required": [
          "project",
          "ok"
        ],
        "title": "ProjectLoadOut",
        "type": "object"
      },
      "ProjectOut": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "captured_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "has_snapshot": {
            "title": "Has Snapshot",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "opcua_url": {
            "title": "Opcua Url",
            "type": "string"
          },
          "snapshot_saved_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Saved At"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "opcua_url",
          "created_at",
          "updated_at",
          "has_snapshot",
          "model_name",
          "model_version",
          "captured_at",
          "snapshot_saved_at",
          "active"
        ],
        "title": "ProjectOut",
        "type": "object"
      },
      "Quality": {
        "description": "Quality of a measured point, collapsed from the OPC UA StatusCode.",
        "enum": [
          "GOOD",
          "UNCERTAIN",
          "BAD",
          "MISSING"
        ],
        "title": "Quality",
        "type": "string"
      },
...
      "Severity": {
        "enum": [
          "info",
          "warning",
          "error"
        ],
        "title": "Severity",
        "type": "string"
      },
...
      "SwitchState": {
        "description": "Position of a switching device.\n\nMaps IEC 61850 Dbpos, but with one addition that matters (AGENTS.md I2):\n`UNDETERMINED` is what we report whenever quality is not GOOD. We never\nturn bad data into a position.",
        "enum": [
          "OPEN",
          "CLOSED",
          "INTERMEDIATE",
          "UNDETERMINED"
        ],
        "title": "SwitchState",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationIssueOut": {
        "description": "A builder issue on the wire, with its UI group (A/B/C).",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "group": {
            "enum": [
              "A",
              "B",
              "C"
            ],
            "title": "Group",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          }
        },
        "required": [
          "severity",
          "code",
          "message",
          "group"
        ],
        "title": "ValidationIssueOut",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Read-only station model from OneATS DataServer",
    "title": "Black Interface — Domain API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/bays/{bay_id}": {
      "get": {
        "operationId": "bay_detail_api_bays__bay_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "bay_id",
            "required": true,
            "schema": {
              "title": "Bay Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BayDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bay Detail"
      }
    },
...
    "/api/login": {
      "post": {
        "description": "Verify a password and start a session.\n\nWrites a row to the local SQLite store and sets a cookie. Nothing here\nreaches OneATS (I1) — this endpoint is about who is looking at the station,\nnot about the station.",
        "operationId": "login_api_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login"
      }
    },
...
    "/api/password": {
      "post": {
        "description": "Change your own password. The current one has to be given again.\n\nOnly your own: changing somebody else's belongs with `account.manage` and\nan audit record, and that screen does not exist yet. Better absent than\nhalf-built.",
        "operationId": "change_password_api_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Password"
      }
    },
    "/api/projects": {
      "get": {
        "operationId": "projects_api_projects_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectOut"
                  },
                  "title": "Response Projects Api Projects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Projects"
      },
      "post": {
        "description": "Create a project and immediately try to browse its DataServer.\n\nSuccess stores a snapshot and makes the project active. Failure keeps the\nproject so the URL can be corrected and retried with /refresh.",
        "operationId": "create_project_api_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectLoadOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project"
      }
    },
    "/api/projects/{project_id}": {
      "delete": {
        "description": "Remove a project and its snapshot. Returns the remaining projects.",
        "operationId": "delete_project_api_projects__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectOut"
                  },
                  "title": "Response Delete Project Api Projects  Project Id  Delete",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Project"
      }
    },
    "/api/projects/{project_id}/open": {
      "post": {
        "description": "Make this project current. Renders from its snapshot when one exists;\nonly its first-ever open touches the DataServer.",
        "operationId": "open_project_api_projects__project_id__open_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectLoadOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Open Project"
      }
    },
    "/api/projects/{project_id}/refresh": {
      "post": {
        "description": "Browse the project's DataServer live and replace its snapshot.",
        "operationId": "refresh_project_api_projects__project_id__refresh_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectLoadOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Project"
      }
    },
...
  }
}